[fix] 서비스에서 조건에 맞지 않으면 예외처리 (#38)
* fix/readme.md

* service checking complete

+++ src/main/java/com/metacoding/springrocketdanv1/board/BoardRepository.java
@@ -26,7 +26,7 @@ public class BoardRepository {
         return em.find(Board.class, id);
     }
 
-    public void deleteById(Long id) {
+    public void deleteById(Integer id) {
         Board board = em.find(Board.class, id);
         if (board != null) {
             em.remove(board);
